httpx to aiohttp and revert #60
