+++ pyproject.toml
@@ -24,7 +24,7 @@ classifiers = [
 dependencies = [
     "aiofiles~=24.1.0",
     "APScheduler~=3.11.0",
-    "cachetools~=5.5.2",
+    "cachetools~=6.0.0",
     "kurigram~=2.2.3",
     "meval~=2.5",
     "ntgcalls~=2.0.0rc7",
