+++ src/helpers/_api.py
@@ -6,14 +6,12 @@ import re
 from pathlib import Path
 from typing import Optional, Union
 
-from pytdbot import types
-
 from src import config
 from src.logger import LOGGER
 from ._dataclass import MusicTrack, PlatformTracks, TrackInfo
 from ._dl_helper import SpotifyDownload
 from ._downloader import MusicService
-from ._httpx import HttpxClient
+from ._aiohttp import AioHttpClient
 
 
 class ApiData(MusicService):
@@ -43,7 +41,7 @@ class ApiData(MusicService):
             query: URL or search query to process
         """
         self.query = self._sanitize_query(query) if query else None
-        self.client = HttpxClient()
+        self.client = AioHttpClient()
         self.api_url = config.API_URL.rstrip("/") if config.API_URL else None
         self.api_key = config.API_KEY
 
@@ -143,7 +141,7 @@ class ApiData(MusicService):
         data = await self._make_api_request("get_track", {"id": self.query})
         return TrackInfo(**data) if data else None
 
-    async def download_track(self, track: TrackInfo, video: bool = False, msg: Union[None, types.Message]= None) -> Optional[Union[str, Path]]:
+    async def download_track(self, track: TrackInfo, video: bool = False) -> Optional[Union[str, Path]]:
         if not track:
             return None
